+++ src/http-server/client/request-parser.js
@@ -17,6 +17,7 @@ export default class VelociousHttpServerClientRequestParser {
     this.configuration = configuration
     this.data = []
     this.events = new EventEmitter()
+    this.hasCompleted = false
     /** @type {Record<string, string | string[] | undefined | Record<string, unknown> | unknown[]>} */
     this.params = {}
 
@@ -54,7 +55,13 @@ export default class VelociousHttpServerClientRequestParser {
    * @param {Buffer} data - Data payload.
    * @returns {void} - No return value.
    */
-  feed = (data) => this.requestBuffer.feed(data)
+  feed = (data) => {
+    if (this.hasCompleted) {
+      throw new Error("Request parser already completed")
+    }
+
+    this.requestBuffer.feed(data)
+  }
 
   /**
    * @param {string} name - Name.
@@ -120,6 +127,7 @@ export default class VelociousHttpServerClientRequestParser {
 
   /** @returns {void} - No return value.  */
   requestDone = () => {
+    this.hasCompleted = true
     incorporate(this.params, this.requestBuffer.params)
 
     this.state = "done"